refactor save share as separate module
