+++ backend/app/core/models.py
@@ -1,0 +1,34 @@
+from datetime import datetime
+from sqlmodel import (
+    Field,
+    SQLModel,
+    UUID,
+    TIMESTAMP,
+    text,
+)
+
+
+class UUIDType(UUID):
+    def __init__(self, *args, **kwargs):
+        kwargs["as_uuid"] = False
+        super().__init__(*args, **kwargs)
+
+
+class TimeStampMixin(SQLModel):
+    created_at: datetime | None = Field(
+        sa_type=TIMESTAMP(timezone=True),
+        sa_column_kwargs={
+            "server_default": text("CURRENT_TIMESTAMP"),
+        },
+        nullable=False,
+        default=None,
+    )
+
+    updated_at: datetime | None = Field(
+        sa_type=TIMESTAMP(timezone=True),
+        sa_column_kwargs={
+            "server_default": text("CURRENT_TIMESTAMP"),
+        },
+        nullable=False,
+        default=None,
+    )
